Do not use first tile as background
This idea will not work as I anticipate a lot of work and it means
writing over the tile map with blanks. It is better to find a way to
overlay text using the hardware or define an upper or lower portion of
the display as a text area (splitting graphics mode), rather than
messing with the tile map.